+++ components/places/patina_places_workspace.sh
@@ -5,7 +5,7 @@
 ###########
 
 # Patina: A 'patina', 'layer', or 'toolbox' for BASH under Linux.
-# Copyright (C) 2020 William Willis Whinn
+# Copyright (C) 2021 William Willis Whinn
 
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -100,7 +100,7 @@ alias 'p-gitupdate'='patina_workspace_update_git_repositories'
 
 # PATINA > FUNCTIONS > FILE MANAGER COMMANDS > WORKSPACE DIRECTORIES
 
-alias 'p-w'='patina_open_folder "$PATINA_PATH_WORKSPACE"'
-alias 'p-w-git'='patina_open_folder "$PATINA_PATH_WORKSPACE_GIT"'
+alias 'workspace'='patina_open_folder "$PATINA_PATH_WORKSPACE"'
+alias 'workspace-git'='patina_open_folder "$PATINA_PATH_WORKSPACE_GIT"'
 
 # End of File.